Remove print state and call print directly (#1872)
This PR calls the print function directly from app.tsx rather than setting up state. By not having to set editor/preview mode we decrease the complexity of this functionality while giving the user the option to print the editor view OR the markdown styled mode.

Before this change, the print function was intended to only print the formatted markdown mode but the functionality is broken. This PR removes the broken code.